move presenter_info into users_schemas
